+++ pkg/scheduler/k8s_internal/predicates/maxNodeResources_test.go
@@ -47,6 +47,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 							v1.ResourceMemory:             resource.MustParse("200Mi"),
 							resource_info.GPUResourceName: resource.MustParse("1"),
 							"kai.scheduler/r1":            resource.MustParse("2"),
+							v1.ResourcePods:               resource.MustParse("110"),
 						}),
 					},
 				},
@@ -84,6 +85,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 							v1.ResourceMemory:             resource.MustParse("200Mi"),
 							resource_info.GPUResourceName: resource.MustParse("1"),
 							"kai.scheduler/r1":            resource.MustParse("2"),
+							v1.ResourcePods:               resource.MustParse("110"),
 						}),
 					},
 					"n2": {
@@ -92,6 +94,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 							v1.ResourceMemory:             resource.MustParse("200Mi"),
 							resource_info.GPUResourceName: resource.MustParse("1"),
 							"kai.scheduler/r1":            resource.MustParse("2"),
+							v1.ResourcePods:               resource.MustParse("110"),
 						}),
 					},
 				},
@@ -117,7 +120,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 			},
 			expected{
 				ksf.NewStatus(ksf.Unschedulable,
-					"The pod n1/name1 requires GPU: 0, CPU: 1 (cores), memory: 0 (GB). Max CPU resources available in a single node in the default node-pool is topped at 0.5 cores"),
+					"The pod n1/name1 requires GPU: 0, CPU: 1 (cores), memory: 0 (GB), pods: 1. Max CPU resources available in a single node in the default node-pool is topped at 0.5 cores"),
 			},
 		},
 		{
@@ -130,6 +133,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 							v1.ResourceMemory:             resource.MustParse("400Mi"),
 							resource_info.GPUResourceName: resource.MustParse("1"),
 							"kai.scheduler/r1":            resource.MustParse("2"),
+							v1.ResourcePods:               resource.MustParse("110"),
 						}),
 					},
 					"n2": {
@@ -138,6 +142,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 							v1.ResourceMemory:             resource.MustParse("200Mi"),
 							resource_info.GPUResourceName: resource.MustParse("1"),
 							"kai.scheduler/r1":            resource.MustParse("2"),
+							v1.ResourcePods:               resource.MustParse("110"),
 						}),
 					},
 				},
@@ -163,7 +168,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 			},
 			expected{
 				ksf.NewStatus(ksf.Unschedulable,
-					"The pod n1/name1 requires GPU: 0, CPU: 0 (cores), memory: 1 (GB). Max memory resources available in a single node in the default node-pool is topped at 0.419 GB"),
+					"The pod n1/name1 requires GPU: 0, CPU: 0 (cores), memory: 1 (GB), pods: 1. Max memory resources available in a single node in the default node-pool is topped at 0.419 GB"),
 			},
 		},
 		{
@@ -176,6 +181,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 							v1.ResourceMemory:             resource.MustParse("200Mi"),
 							resource_info.GPUResourceName: resource.MustParse("1"),
 							"kai.scheduler/r1":            resource.MustParse("2"),
+							v1.ResourcePods:               resource.MustParse("110"),
 						}),
 					},
 					"n2": {
@@ -184,6 +190,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 							v1.ResourceMemory:             resource.MustParse("200Mi"),
 							resource_info.GPUResourceName: resource.MustParse("1"),
 							"kai.scheduler/r1":            resource.MustParse("2"),
+							v1.ResourcePods:               resource.MustParse("110"),
 						}),
 					},
 				},
@@ -209,7 +216,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 			},
 			expected{
 				ksf.NewStatus(ksf.Unschedulable,
-					"The pod n1/name1 requires GPU: 2, CPU: 0 (cores), memory: 0 (GB). Max GPU resources available in a single node in the default node-pool is topped at 1"),
+					"The pod n1/name1 requires GPU: 2, CPU: 0 (cores), memory: 0 (GB), pods: 1. Max GPU resources available in a single node in the default node-pool is topped at 1"),
 			},
 		},
 		{
@@ -222,6 +229,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 							v1.ResourceMemory:             resource.MustParse("200Mi"),
 							resource_info.GPUResourceName: resource.MustParse("0"),
 							"kai.scheduler/r1":            resource.MustParse("2"),
+							v1.ResourcePods:               resource.MustParse("110"),
 						}),
 					},
 					"n2": {
@@ -230,6 +238,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 							v1.ResourceMemory:             resource.MustParse("200Mi"),
 							resource_info.GPUResourceName: resource.MustParse("0"),
 							"kai.scheduler/r1":            resource.MustParse("2"),
+							v1.ResourcePods:               resource.MustParse("110"),
 						}),
 					},
 				},
@@ -254,7 +263,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 			},
 			expected{
 				ksf.NewStatus(ksf.Unschedulable,
-					"The pod n1/name1 requires GPU: 0.5, CPU: 0 (cores), memory: 0 (GB). No node in the default node-pool has GPU resources"),
+					"The pod n1/name1 requires GPU: 0.5, CPU: 0 (cores), memory: 0 (GB), pods: 1. No node in the default node-pool has GPU resources"),
 			},
 		},
 		{
@@ -268,6 +277,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 							resource_info.GPUResourceName: resource.MustParse("1"),
 							"kai.scheduler/r1":            resource.MustParse("2"),
 							v1.ResourceEphemeralStorage:   resource.MustParse("10Gi"),
+							v1.ResourcePods:               resource.MustParse("110"),
 						}),
 					},
 					"n2": {
@@ -277,6 +287,7 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 							resource_info.GPUResourceName: resource.MustParse("1"),
 							"kai.scheduler/r1":            resource.MustParse("2"),
 							v1.ResourceEphemeralStorage:   resource.MustParse("20Gi"),
+							v1.ResourcePods:               resource.MustParse("110"),
 						}),
 					},
 				},
@@ -302,13 +313,18 @@ func Test_podToMaxNodeResourcesFiltering(t *testing.T) {
 			},
 			expected{
 				ksf.NewStatus(ksf.Unschedulable,
-					"The pod n1/name1 requires GPU: 0, CPU: 0 (cores), memory: 0 (GB), ephemeral-storage: 25 (GB). "+
+					"The pod n1/name1 requires GPU: 0, CPU: 0 (cores), memory: 0 (GB), ephemeral-storage: 25 (GB), pods: 1. "+
 						"Max ephemeral-storage resources available in a single node in the default node-pool is topped at 21.474 GB"),
 			},
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			for _, node := range tt.args.nodesMap {
+				if _, found := node.Allocatable.ScalarResources()[v1.ResourcePods]; !found {
+					node.Allocatable.ScalarResources()[v1.ResourcePods] = 110
+				}
+			}
 			mnr := NewMaxNodeResourcesPredicate(tt.args.nodesMap, tt.args.resourceClaims, tt.args.nodePoolName)
 			_, status := mnr.PreFilter(context.TODO(), nil, tt.args.pod, nil)
 			if !statusEqual(status, tt.expected.status) {
@@ -392,11 +408,13 @@ func TestMaxNodeResourcesPredicateDRA(t *testing.T) {
 							v1.ResourceCPU:     resource.MustParse("100m"),
 							v1.ResourceMemory:  resource.MustParse("200Mi"),
 							"kai.scheduler/r1": resource.MustParse("2"),
+							"pods":             resource.MustParse("110"),
 						},
 						"n2": {
 							v1.ResourceCPU:     resource.MustParse("500m"),
 							v1.ResourceMemory:  resource.MustParse("200Mi"),
 							"kai.scheduler/r1": resource.MustParse("2"),
+							"pods":             resource.MustParse("110"),
 						},
 					},
 				),
@@ -457,11 +475,13 @@ func TestMaxNodeResourcesPredicateDRA(t *testing.T) {
 							v1.ResourceCPU:     resource.MustParse("100m"),
 							v1.ResourceMemory:  resource.MustParse("200Mi"),
 							"kai.scheduler/r1": resource.MustParse("2"),
+							"pods":             resource.MustParse("110"),
 						},
 						"n2": {
 							v1.ResourceCPU:     resource.MustParse("500m"),
 							v1.ResourceMemory:  resource.MustParse("200Mi"),
 							"kai.scheduler/r1": resource.MustParse("2"),
+							"pods":             resource.MustParse("110"),
 						},
 					},
 				),
@@ -508,7 +528,7 @@ func TestMaxNodeResourcesPredicateDRA(t *testing.T) {
 				},
 			},
 			ksf.NewStatus(ksf.Unschedulable,
-				"The pod n1/name1 requires GPU: 2, CPU: 0 (cores), memory: 0 (GB). Max GPU resources available in a single node in the default node-pool is topped at 1"),
+				"The pod n1/name1 requires GPU: 2, CPU: 0 (cores), memory: 0 (GB), pods: 1. Max GPU resources available in a single node in the default node-pool is topped at 1"),
 		},
 	}
 	for _, tt := range tests {